refactor(raijin): colocate invocation interfaces
