+++ yarn/raijin/src/commands/invocation/adapters/path/path.interfaces.ts
@@ -1,0 +1,5 @@
+import type { PortablePath } from '@yarnpkg/fslib'
+
+export interface PathAdapter {
+  toNative: (cwd: PortablePath) => string
+}
