+++ yarn/raijin/src/commands/invocation/index.ts
@@ -9,7 +9,7 @@ export { shouldProxyCommand }         from './proxy.js'
 export { resolveProjectInvocation }   from './resolve.js'
 export { resolveWorkspaceInvocation } from './resolve.js'
 
-export type { YarnExecutable }        from './adapters/yarn/execution.js'
-export type { YarnExecutableOptions } from './adapters/yarn/execution.js'
-export type { ProjectInvocation }     from './resolve.js'
-export type { WorkspaceInvocation }   from './resolve.js'
+export type { YarnExecutable }        from './adapters/yarn/execution.interfaces.js'
+export type { YarnExecutableOptions } from './adapters/yarn/execution.interfaces.js'
+export type { ProjectInvocation }     from './resolve.interfaces.js'
+export type { WorkspaceInvocation }   from './resolve.interfaces.js'
